review: split test files
