+++ testdata/src/excluded/binary.go
@@ -1,0 +1,11 @@
+package excluded
+
+type Binary struct{}
+
+func (b Binary) MarshalBinary() ([]byte, error) {
+	panic("not implemented")
+}
+
+func (b *Binary) UnmarshalBinary(data []byte) error {
+	panic("not implemented")
+}
