+++ testdata/src/excluded/gob.go
@@ -1,0 +1,11 @@
+package excluded
+
+type Gob struct{}
+
+func (g Gob) GobEncode() ([]byte, error) {
+	panic("not implemented")
+}
+
+func (g *Gob) GobDecode(data []byte) error {
+	panic("not implemented")
+}
