+++ testdata/src/excluded/json.go
@@ -1,0 +1,11 @@
+package excluded
+
+type JSON struct{}
+
+func (j JSON) MarshalJSON() ([]byte, error) {
+	panic("not implemented")
+}
+
+func (j *JSON) UnmarshalJSON(b []byte) error {
+	panic("not implemented")
+}
