+++ testdata/src/excluded/text.go
@@ -1,0 +1,11 @@
+package excluded
+
+type Text struct{}
+
+func (t Text) MarshalText() ([]byte, error) {
+	panic("not implemented")
+}
+
+func (t *Text) UnmarshalText(b []byte) error {
+	panic("not implemented")
+}
