+++ testdata/src/excluded/xml.go
@@ -1,0 +1,13 @@
+package excluded
+
+import "encoding/xml"
+
+type XML struct{}
+
+func (x XML) MarshalXML(e *xml.Encoder, start xml.StartElement) error {
+	panic("not implemented")
+}
+
+func (x *XML) UnmarshalXML(d *xml.Decoder, start xml.StartElement) error {
+	panic("not implemented")
+}
